+++ lib/app/common/app_page/app_page.dart
@@ -27,6 +27,7 @@ import 'package:software/app/common/app_page/app_swipe_gesture.dart';
 import 'package:software/app/common/app_page/media_tile.dart';
 import 'package:software/app/common/app_page/page_layouts.dart';
 import 'package:software/app/common/border_container.dart';
+import 'package:software/app/common/custom_back_button.dart';
 import 'package:software/app/common/safe_network_image.dart';
 import 'package:software/l10n/l10n.dart';
 import 'package:yaru_icons/yaru_icons.dart';
@@ -263,7 +264,7 @@ class _AppPageState extends State<AppPage> {
                   icon: const Icon(YaruIcons.document_open),
                 ),
               )
-            : _CustomBackButton(
+            : CustomBackButton(
                 onPressed: () => Navigator.pop(context),
               ),
       ),
@@ -344,29 +345,6 @@ class _CarouselDialogState extends State<_CarouselDialog> {
             ),
           )
         ],
-      ),
-    );
-  }
-}
-
-class _CustomBackButton extends StatelessWidget {
-  const _CustomBackButton({
-    Key? key,
-    this.onPressed,
-  }) : super(key: key);
-
-  final Function()? onPressed;
-
-  @override
-  Widget build(BuildContext context) {
-    return Center(
-      child: IconButton(
-        style: IconButton.styleFrom(fixedSize: const Size(40, 40)),
-        onPressed: () {
-          Navigator.maybePop(context);
-          if (onPressed != null) onPressed!();
-        },
-        icon: const Icon(YaruIcons.go_previous),
       ),
     );
   }
