+++ lib/app/common/custom_back_button.dart
@@ -1,0 +1,39 @@
+/*
+ * Copyright (C) 2022 Canonical Ltd
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License version 3 as
+ * published by the Free Software Foundation.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ *
+ */
+
+import 'package:flutter/material.dart';
+import 'package:yaru_icons/yaru_icons.dart';
+
+class CustomBackButton extends StatelessWidget {
+  const CustomBackButton({
+    Key? key,
+    this.onPressed,
+  }) : super(key: key);
+
+  final Function()? onPressed;
+
+  @override
+  Widget build(BuildContext context) {
+    return Center(
+      child: IconButton(
+        style: IconButton.styleFrom(fixedSize: const Size(40, 40)),
+        onPressed: onPressed ?? () => Navigator.maybePop(context),
+        icon: const Icon(YaruIcons.go_previous),
+      ),
+    );
+  }
+}
